refac: ui part를 상속받아 구현하도록 변경
